+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Changed
 - practicalli: revert to upstream other.nvim
 - community: add rainbow-delimiters with theme support
+- community: add color picker ccc-nvim
 
 # 2025-04-24